Deleted OffCanvasExample component, added it to Navbar
