+++ src/Styles/navbar_styles.scss
@@ -394,6 +394,10 @@
 //   }
 // }
 
+.offcanvas-debug {
+  border: 2px solid red;
+}
+
